# compilation_service/swagger.yaml
openapi: 3.0.3
info:
  title: Compile Service API
  version: 1.0.0
  description: API для компиляции кода с разными типами входных данных
paths:
  /metrics:
    get:
      summary: Получить метрики Prometheus
      description: Возвращает метрики в формате Prometheus.
      responses:
        '200':
          description: Метрики в формате Prometheus
          content:
            text/plain; version=0.0.4; charset=utf-8:
              schema:
                type: string
                example: '# HELP http_requests_total The total number of HTTP requests.\n# TYPE http_requests_total counter\nhttp_requests_total{method="get",code="200"} 1027 1395066363000\n'

  /api/compile:
    post:
      summary: Компиляция кода из строки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                
              properties:
                code:
                  type: string
                  description: Исходный код для компиляции
                input:
                  type: string
                  description: Входные данные для программы

                filename:
                  type: string
                  description: Имя файла с входными данными для программы

                timeout:
                  type: number
                  description: Ограничение по времени для исполнения кода

      responses:
        '200':
          description: Результат компиляции
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Выходные данные программы
                  error:
                    type: string
                    description: Сообщение об ошибке (если есть)
        '500':
          description: Ошибка компиляции
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Сообщение об ошибке
        '401':
          description: Неавторизованный доступ
  /api/compile/file:
    post:
      summary: Компиляция из исполняемого файла
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Исполняемый файл с кодом
                input:
                  type: string
                  description: Входные данные для программы
                inputFilename:
                  type: string
                  description: Имя файла с входными данными
                timeout:
                  type: number
                  description: Ограничение по времени для исполнения кода
                  
      responses:
        '200':
          description: Результат компиляции
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Выходные данные программы
                  error:
                    type: string
                    description: Сообщение об ошибке (если есть)
        
        '401':
          description: Неавторизованный доступ

        '500':
          description: Ошибка компиляции
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Сообщение об ошибке
  /api/compile/archive:
    post:
      summary: Компиляция кода из архива
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - archive
                - type
              properties:
                archive:
                  type: string
                  format: binary
                  description: Архив с исходным кодом (zip)
                input:
                  type: string
                  description: Входные данные для программы
                type:
                  type: string
                  enum:
                    - stdin
                    - file
                  description: Способ передачи входных данных
      responses:
        '200':
          description: Результат компиляции
          content:
            application/json:
                          schema:
                            type: object
                            properties:
                              output:
                                type: string
                                description: Выходные данные программы
                              error:
                                type: string
                                description: Сообщение об ошибке (если есть)
        '401':
          description: Неавторизованный доступ
components: {}